refactor: add helpful utils functions to help track y motion
In this commit, we add utilites that determine if the player is standing, some helpers to track the sprite tile you are currently standing in, as well as function that zips you up to the nearest top of a tile to stand on.

This latter 'zipping' function is incomplete, and will not ensure that you reach a standing position.