refactor(colors): replace the hard-coded colors with per-layer color mapping
This commit makes drastic strides in favour of simplifying the way we work with colors in the game engine.

Before this commit, we were  hard-coding a static list of colors, then defining sprites as an array of numbers that corresponded to indicies in that list. The problem with that method is that it is difficult to change colors, and their position, once they are defined.

This new method splits the list into two pieces:
- greyscale colors: a static list of greyscale colors only
- layer-based color mappings: there is currently a player colormap and a level colormap. It's much easier now to start designing characters from the greyscale level, then put on the color theorist hat on to add coehrent a color palette mapping for each layer.
  - eg. we may want to have the player stand out, so we can create a suite of colors that do just that. :)