feat: calculate approx FPS and display
Strangely, there is a bug where if you navigate to another tab within the same browser as the game and tab back, the next time the FPS number updates, the number will be really big.

This is unexpected behaviour and will need to be addressed in a future PR or commit.