build: allow support for implicit color map
More specifically, this commit makes sure that you are allowed to generate sprite data with the Python utility script without feeding in a preset mapping from colors. When not provided, the function will just use the colors in the PNG image itself and make the mapping from there.

This commit also adds a print statement when you run the main script command 'make sprites' which will print that new mapping if generated.