refactor(sprites): attach padding data to sprites
This commit attaches pre-processed data for each protagonist that tells you the number of padding (in units relative to the sprite array) of blank space on all 4 sides.

This change was motivated by the fact that the player is typically a little thinner than the 16x16 grid spcae it is defined in, and as such when we want to create a boundary system in the game, it is important to know not just the 16x16 space, but the TRUE space the player occupies.