feat: add item-collecting mechanic
This commit swaps the big level for a smaller test one, adds a new sprite with name "gold", and places gold around the level.

In addition to that, a new mechanicsm is added to allow the player to run into into the gold, making it disappear, and add to the total gold.