feat(editor): save level so compatible w/ engine
This commit exports not only the level array, but a level array for the enemies and the items when you press the "Copy to Clipboard" button. These last 2 arrays are empty, but are exported as the code in the engine will break if these components of a level are missing.