feat: increase spawn speed
We need to make the speed more managable. There was also a game over state added to the STATE constant, but it is only really a placeholder.

There is a helper function that tells us the number of slots that are open; we can use this in the future to determine the pace of the spawner mechanism.