feat: correct spawn interval to make game fun
The game is now really fun immediately. Now, we gradually increase the spawn interval of new pieces once every 10 points are gained. As an emergent gameplay mechanic, placing a shape on an identical shape will not increase the points at all. This is neutral and can be used as a tactic when the pieces are filling in very very quickly.