feat(text): prevent user selection
This commit prevents the user from selecting the text contained in tags with class 'text'. Previously, I had thought that setting the pointer-events to none would do the trick, but this did not: I believe now that a user highlighting text, for example, is not a pointer event, but clicking is a pointer event.