fix(tutorial): prevent twitchy behaviour
Previously, when you ran your cursor into the wall, the text above the cursor (called the 'tutorial' in the code) would fidget in place. The reason is because the position of the tutorial was being assigned before the boundary check was being ran. The boundary check, resulting in an immediate relocation on the next frame.